requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
phonenumbers==8.13.30
pillow==10.2.0
psycopg2-binary==2.9.9
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
